pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "1.0.0"
description = "FastAPI Boilerplate - With RBAC Authentication System"
authors = ["Amit Dhakal <herojk64@gmail.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.pyright]
include = ["."]
venvPath = "."
venv = "env"
pythonVersion = "3.11"

[project]
name = "fastapi-boilerplate"
version = "1.0.0"
description = "FastAPI Boilerplate - RBAC Authentication System"
authors = [
    {name = "Amit Dhakal",email = "herojk64@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "alembic (==1.17.2)",
    "annotated-doc (==0.0.4)",
    "annotated-types (==0.7.0)",
    "anyio (==4.12.0)",
    "asyncio (==4.0.0)",
    "asyncpg (==0.31.0)",
    "bcrypt (==5.0.0)",
    "certifi (==2025.11.12)",
    "cffi (==2.0.0)",
    "click (==8.3.1)",
    "cryptography (==46.0.3)",
    "dnspython (==2.8.0)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.3.0)",
    "fastapi (==0.123.5)",
    "fastapi-cli (==0.0.16)",
    "fastapi-cloud-cli (==0.5.2)",
    "fastar (==0.8.0)",
    "greenlet (==3.3.0)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httptools (==0.7.1)",
    "httpx (==0.28.1)",
    "idna (==3.11)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.6)",
    "mako (==1.3.10)",
    "markdown-it-py (==4.0.0)",
    "markupsafe (==3.0.3)",
    "mdurl (==0.1.2)",
    "orjson (==3.11.4)",
    "passlib (==1.7.4)",
    "psycopg2-binary (==2.9.11)",
    "pyasn1 (==0.6.1)",
    "pycparser (==2.23)",
    "pydantic (==2.12.5)",
    "pydantic-extra-types (==2.10.6)",
    "pydantic-settings (==2.12.0)",
    "pydantic-core (==2.41.5)",
    "pygments (==2.19.2)",
    "python-dotenv (==1.2.1)",
    "python-jose (==3.5.0)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.3)",
    "rich (==14.2.0)",
    "rich-toolkit (==0.17.0)",
    "rignore (==0.7.6)",
    "rsa (==4.9.1)",
    "sentry-sdk (==2.46.0)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sqlalchemy (==2.0.44)",
    "starlette (==0.50.0)",
    "typer (==0.20.0)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "ujson (==5.11.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.38.0)",
    "uvloop (==0.22.1)",
    "watchfiles (==1.1.1)",
    "websockets (==15.0.1)",
    "pytest (>=9.0.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "requests (>=2.32.5,<3.0.0)"
]
